// gui/src/rb/tokens.d.ts
export type RbThemeName = 'light' | 'dark';
// ...
export declare const rounded: {
    readonly full: 9999;
    readonly "4xl": 32;
    readonly "3xl": 24;
    readonly "2xl": 16;
    readonly xl: 12;
    readonly sm: 4;
    readonly lg: 8;
    readonly md: 6;
    readonly xs: 2;
};